requirements.txt
Pillow~=9.5.0
cython
matplotlib~=3.5.3
scikit-image
opencv-python~=4.7.0.72
h5py
imgaug
pandas~=1.1.5
argparse
IPython[all]
simplejson~=3.19.1
sklearn~=0.0.post1
slidingwindow
pyyaml~=6.0
numpy~=1.19.5
torch~=1.9.1+cu111
pyqt5~=5.15.9
scikit-learn~=0.21.3
scipy~=1.7.3
tensorflow-gpu~=2.4.0
seaborn~=0.12.2
torchvision~=0.10.1+cu111
tqdm~=4.65.0
thop~=0.1.1-2209072238
requests~=2.28.2